Restructured the components
