+++ templates/stride/components/elements/HorizontalRuler.tsx
@@ -1,6 +1,6 @@
 import { ContentProps, contentType } from '@optimizely/cms-sdk';
 import { getPreviewUtils } from '@optimizely/cms-sdk/react/server';
-import EditableField from '../EditableField';
+import EditableField from '../shared/EditableField';
 
 export const HorizontalRulerComponent = contentType({
   key: 'HorizontalRulerElement',
