+++ templates/stride/components/elements/Video.tsx
@@ -1,6 +1,6 @@
 import { ContentProps, contentType } from '@optimizely/cms-sdk';
 import { getPreviewUtils } from '@optimizely/cms-sdk/react/server';
-import EditableField from '../EditableField';
+import EditableField from '../shared/EditableField';
 
 export const VideoComponent = contentType({
   key: 'VideoElement',
