+++ templates/stride/components/layouts/shared/DecoGrid.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '../../lib/utils';
+import { cn } from '../../../lib/utils';
 
 interface CtaProps {
   children?: React.ReactNode;
