+++ templates/stride/components/layouts/shared/Footer.tsx
