+++ templates/stride/components/layouts/shared/Header.tsx
@@ -1,5 +1,5 @@
-import { HeaderClient } from './HeaderClient';
-import { getNavigationItems } from '../../lib/navigation';
+import { HeaderClient } from './Header.client';
+import { getNavigationItems } from '../../../lib/navigation';
 
 export const Header = async () => {
   const navigationItems = await getNavigationItems();
