+++ templates/stride/components/layouts/shared/Logo.tsx
