+++ templates/stride/components/layouts/shared/MobileMenu.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { X, ChevronRight, ChevronLeft } from 'lucide-react';
-import { cn } from '../../lib/utils';
+import { cn } from '../../../lib/utils';
 import { Logo } from './Logo';
 
 interface MobileMenuProps {
