+++ templates/stride/components/layouts/shared/SearchModal.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Search, X, Calendar, MapPin, FileText, ExternalLink } from 'lucide-react';
 import Link from 'next/link';
-import { cn } from '../../lib/utils';
+import { cn } from '../../../lib/utils';
 
 interface SearchModalProps {
   isOpen: boolean;
