+++ templates/stride/components/pages/Start.tsx
@@ -1,6 +1,6 @@
 import { contentType, ContentProps } from '@optimizely/cms-sdk';
 import { OptimizelyComposition } from '@optimizely/cms-sdk/react/server';
-import FullWidthLayout from '../layout/FullWidthLayout';
+import FullWidthLayout from '../layouts/FullWidthLayout';
 
 export const StartPage = contentType({
   key: 'StartPage',
