+++ templates/stride/components/shared/EditableField.tsx
